Fix shaping problems on iOS 14 / macOS 11
Detection of AAT fonts was exclusively based on the existence
of tables mort, morx and kerx. In the SF font (the UI default)
on iOS 14 and Big Sur, none of these tables actually exist, so
we would accidentally use the OpenType shaper for it. But there
are plenty of other AAT tables in the font and the OT shaper
can´t handle it. In particular, this caused us to get the wrong
advance for comma when it was followed by a space.

Looking at all the affected SF fonts on Big Sur, they all have
the trak table, so we expand the detection to include this.
This is possibly also the table used to determine the correct
glyph positions in this case.

In Qt 6, this already works as expected, because Harfbuzz has
built-in support for AAT and does not require the CoreText
fallback. Therefore, this is just a hotfix for the 5.15 series
and there is no need to upstream this patch.

[ChangeLog][macOS] Fixed shaping of default UI font on macOS 11
and iOS 14.

Fixes: QTBUG-88495
Change-Id: Ia2b3f45cbc0c32b93864760cb427ede6b0566e8f
Reviewed-by: Tor Arne Vestbø <tor.arne.vestbo@qt.io>
Reviewed-by: Lars Knoll <lars.knoll@qt.io>