macOS: Account for Big Sur always enabling layer backed views
Previous versions of macOS (Mojave and Catalina) enabled layer backing
only when the application's build SDK was 10.14 or above, to provide
an opt-out workaround for applications that were not ready for layer
backing. This seems to no longer be the case, as Big Sur will enable
layer backing regardless of the SDK version used.

Tested by passing -Wl,-sdk_version -Wl,10.13 -v -mlinker-version=519
during linking, to simulate what the effects of linking to the 10.13
would be. Also confirmed by real-world use of the 10.13 SDK via Xcode
9.

Change-Id: I867edc001b6709ba0ad0fd70918bd9b77d317cf8
Fixes: QTBUG-87014
Pick-to: 5.12
Reviewed-by: Alexandru Croitor <alexandru.croitor@qt.io>
Reviewed-by: Timur Pocheptsov <timur.pocheptsov@qt.io>