Be more consistent when converting JSON values from variant
When converting from numeric QVariant types to QJsonValue, we always
convert to double. However, when converting from QVariantMap or
QVariantList, we convert to qint64, which may result to negative number
in the corner cases. Fixed to always cast to double, which matches with
the pre-5.15.0 behavior. Note, that in Qt 6 QJsonValue got native
support for integers, and we consistently fallback to double only if the
value is outside the range of qint64, so this fix applies only to 5.15.

[ChangeLog][Important Behavior Changes] Restored pre-5.15.0 behavior
when converting from numeric QVariant values to QJson* types. Such
values now always convert to a double QJsonValue.

Fixes: QTBUG-88168
Change-Id: I584a35aa6ffade22a1c83fcda5598ed912f53919
Reviewed-by: Thiago Macieira <thiago.macieira@intel.com>
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>