QMutex: Work around ICC bug in dealing with constexpr constructor
The bug report says it also happens in some MSVC builds, but no one has
so far been able to reproduce it or give us a log with more details, so
not working around the issue for MSVC. ICC, howevere, is known to have a
lot of parsing issues, so I'm not surprised and don't feel a problem
working around them.

This fix is not applicable for Qt 6.

Fixes: QTBUG-85644
Change-Id: If51855da004b4f3fbf43fffd1649b64e05cb91cc
Reviewed-by: Allan Sandfeld Jensen <allan.jensen@qt.io>