CMake: Use lower cased property name for entrypoint library opt-out
We need this to enable headersclean functionality to work with CMake
<=3.18 which is our minimum required CMake version to build
Qt (it's not a problem for 3.19).

Otherwise we hit a configuration error while trying to evaluate
generator expressions for INTERFACE only libraries for which we need
to do the headersclean check (e.g. UiPlugin in qttools).

Sample error message:

 INTERFACE_LIBRARY targets may only have whitelisted properties.  The
  property "QT_NO_ENTRYPOINT" is not allowed.

More info:

 https://gitlab.kitware.com/cmake/cmake/-/issues/21484

Naming it lower case avoids the issue, but is inconsistent with all
our other public property names.

Furthermore if we do this change, it should probably go to 6.0.0 due
to it being a property that a user is meant to set in their project.

Task-number: QTBUG-82615
Pick-to: 6.0 dev
Change-Id: I1f1439f1e5cbc816010eba7301605a8f60b88fd1
Reviewed-by: Alexandru Croitor <alexandru.croitor@qt.io>